Serializer will return null when id is undefined
Serializer used to return null when there were no attributes.
This caused problems for creating models that only had relationships.

I changed the serializer to return null when id is undefined.
This broke include tests because they expected customMethod to
be able to return attributes without an id.

JSONAPI require id to be present in a resource document,
but the id can be null.

Note: This commit changes expectation of customMethod to return an id, which can be null.