feat(serializer): support multi-level includes
BREAKING CHANGE: remove "requestedIncludes" option
This option should not handle the way this adapter serialize the resources, if there is embedded resources, even if they do not appear in this option, recommended behavior is to have them in the include property as well.
Security consideration: nested includes were previously not handled, it is now the case, like in the loopback documentation, you need to ensure it is secured your side.